feat: stub ChainGetEvents
